Migrated to cobra. start [server,client], completion, run, ...
